Update Nomor 5.py

+++ Nomor 5.py
@@ -1,6 +1,6 @@
 # bab1_Projek_python_nomor_5
 # Meminta input jari-jari lingkaran dari pengguna
-jari_jari = float(input("Masukkan jari-jari lingkaran: "))
+jari_jari = int(input("Masukkan jari-jari lingkaran: "))
 
 # Menghitung luas lingkaran
 luas = 3.14159 * jari_jari**2
